created a BoardsList.js file to remose <Select> from App.js
