+++ src/components/Board.js
@@ -2,6 +2,8 @@
 // import PropTypes from "prop-types";
 // import Cards from "./components";
 
+// RENDERS STATE OF CARDS BASED ON BOARD ID
+
 // const Board = (props) => {
 //   // - does it need to create state for each new board?
 //   // - do we put each card in the same state and pull based on board id?
@@ -10,6 +12,7 @@
 //   // --> GET request from back-end API
 
 //   return <div>~*~Rendering Board.js~*~</div>;
+//      --> <CreateCard> renders HERE! <--
 // };
 
 // export default Board;
